[HW12] Reorganized app components
