+++ hw12_13_14_15_calendar/go.mod
@@ -3,6 +3,7 @@ module github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar
 go 1.16
 
 require (
+	github.com/sirupsen/logrus v1.8.1 // indirect
 	github.com/spf13/cobra v1.3.0
 	github.com/spf13/viper v1.10.1
 )
