+++ hw12_13_14_15_calendar/internal/logger/logger.go
@@ -1,20 +1,17 @@
 package logger
 
-import "fmt"
+import (
+	"os"
 
-type Logger struct { // TODO
+	"github.com/sirupsen/logrus"
+)
+
+func New(level string) *logrus.Logger {
+	log := logrus.New()
+	log.Out = os.Stdout
+	log.Level = logrus.DebugLevel
+	if lvl, err := logrus.ParseLevel(level); err != nil {
+		log.Level = lvl
+	}
+	return log
 }
-
-func New(level string) *Logger {
-	return &Logger{}
-}
-
-func (l Logger) Info(msg string) {
-	fmt.Println(msg)
-}
-
-func (l Logger) Error(msg string) {
-	// TODO
-}
-
-// TODO
