+++ hw12_13_14_15_calendar/run/cmd/version.go
@@ -1,4 +1,4 @@
-package cmd
+package run
 
 import (
 	"encoding/json"
