+++ hw12_13_14_15_calendar/run/main.go
@@ -5,5 +5,5 @@ import (
 )
 
 func main() {
-	cmd.Execute()
+	run.Execute()
 }
